// src/auth/interface/user.interface.ts
export enum UserRole {
    PATIENT = 'patient',
    DOCTOR = 'doctor',
    ADMIN = 'admin',
  }
  
  
  export enum UserStatus {
    ACTIVE = 'active',
    INACTIVE = 'inactive',
    SUSPENDED = 'suspended',
  }


  export enum UserRoomRole {
    PARTICIPANT = 'participant',
    MODERATOR = 'moderator',
    OWNER = 'owner',
  }
  